fix: receive data cause login process failed

+++ Noodoe assignment/Model/UserData.swift
@@ -32,6 +32,6 @@ import Foundation
 struct UserData: Codable {
     let username: String
     let sessionToken: String
-    let timezone: Int
+    let timezone: Double
     let objectId: String
 }
